Extracting rpm-mappings from pdc
+ added tests
